tourism/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart  
aiosqlite
fastapi-cache
redis
greenlet